library_management_system/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
asgiref==3.8.1
astroid==3.2.4
attrs==24.2.0
azure-ai-ml==1.19.0
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-monitor-opentelemetry-exporter==1.0.0b28
azure-storage-blob==12.22.0
azure-storage-file-datalake==12.16.0
azure-storage-file-share==12.17.0
billiard==4.2.1
black==24.8.0
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==43.0.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.7
django-cache-url==3.4.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring_parser==0.16
drf-registration==0.1.3
drf-yasg==1.21.7
environs==11.0.0
fastapi==0.112.2
filelock==3.15.4
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google_search_results==2.4.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.4
inflection==0.5.1
isodate==0.6.1
isort==5.13.2
itsdangerous==2.2.0
itypes==1.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kombu==5.4.2
MarkupSafe==2.1.5
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mccabe==0.7.0
more-itertools==10.4.0
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
mysqlx-connector-python==9.0.0
numpy==2.1.0
oauthlib==3.2.2
openai==1.42.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pbr==6.0.0
pillow==10.4.0
platformdirs==3.0.0
portalocker==2.10.1
prompt_toolkit==3.0.48
promptflow==1.15.0
promptflow-azure==1.15.0
promptflow-core==1.15.0
promptflow-devkit==1.15.0
promptflow-tools==1.4.0
promptflow-tracing==1.15.0
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydash==7.0.7
pygitguardian==1.17.0
Pygments==2.18.0
PyJWT==2.6.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.2
redis==5.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==12.5.1
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==72.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
starlette==0.38.2
stevedore==5.2.0
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.7.0
tomlkit==0.13.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
waitress==2.1.2
wcwidth==0.2.13
Werkzeug==3.0.4
whitenoise==6.7.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.1
zstd==1.5.5.1
